requirements.txt
asgiref==3.8.1
Django==5.1.1
django-redis==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.8
sqlparse==0.5.1
